Change channel/role position and bitrate type to u32
Position type Was inconsistent before; either i32 or u64. I tested a bit and negative numbers aren't really supported (they outright fail in channel position, and role position edit doesn't do anything at all however I use it... but the position numbers in role position edit seem to be relative to each other anyways so it doesn't matter which integers they are)

+++ src/model/guild/member.rs
@@ -297,7 +297,7 @@ impl Member {
     /// position. If two or more roles have the same highest position, then the
     /// role with the lowest ID is the highest.
     #[cfg(feature = "cache")]
-    pub fn highest_role_info(&self, cache: impl AsRef<Cache>) -> Option<(RoleId, i64)> {
+    pub fn highest_role_info(&self, cache: impl AsRef<Cache>) -> Option<(RoleId, u32)> {
         let guild = cache.as_ref().guild(self.guild_id)?;
 
         let mut highest = None;
